+++ src/begin/userscript_definition.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name	Twitter - Mark Last Read
-// @version 1.3.3
+// @version 1.3.4
 // @grant   none
 // @include https://*twitter.com/*
 // ==/UserScript==